♻️refactor : 유저 식별자 email로 변경
