+++ noti-service/src/main/java/com/waither/notiservice/domain/FireBaseToken.java
@@ -11,7 +11,7 @@ import org.springframework.data.redis.core.RedisHash;
 public class FireBaseToken {
 
     @Id
-    private Long userId;
+    private String email;
 
     private String token;
 }
