+++ noti-service/src/main/java/com/waither/notiservice/domain/Notification.java
@@ -20,6 +20,6 @@ public class Notification extends BaseEntity {
 
     private String content;
 
-    private Long userId;
+    private String email;
 
 }
